refactor(engine): classify AI proposal outcomes
